[drawer] Add virtual keyboard provider
