+++ docs/src/app/(docs)/react/components/drawer/page.mdx
@@ -46,7 +46,7 @@ import { Drawer } from '@base-ui/react/drawer';
 
 Drawer supports swipe gestures to dismiss. Set `swipeDirection` to control which direction dismisses the drawer. `<Drawer.Content>` allows text selection of its children without swipe interference when using a mouse pointer. Add `data-base-ui-swipe-ignore` to a descendant when you need to opt that element out of swipe dismissal for all input types.
 
-Use `virtualKeyboardAware` when a drawer contains form fields and you want Base UI to manage visible viewport sizing and keyboard-aware focus handling for software keyboards. Existing drawers keep the legacy behavior by default.
+Use `<Drawer.VirtualKeyboardProvider>` when a drawer contains form fields and you want Base UI to manage visible viewport sizing and keyboard-aware focus handling for software keyboards. Existing drawers keep the legacy behavior by default.
 
 ## Examples
 
@@ -143,10 +143,12 @@ By default, the drawer can skip snap points when swiping quickly. Specify the `s
 
 ### Virtual keyboard aware
 
-Set `virtualKeyboardAware` when a bottom-sheet drawer contains form controls and you want the drawer to react to software keyboards. This pattern keeps the main list scrollable while a footer field stays pinned outside the scroll container.
+Wrap a bottom-sheet drawer in `<Drawer.VirtualKeyboardProvider>` when it contains form controls and you want the drawer to react to software keyboards. This pattern keeps the main list scrollable while a footer field stays pinned outside the scroll container. Use `--available-height` to size the popup against the reduced visual viewport and `--drawer-keyboard-inset` to lift the sheet when the keyboard overlaps the bottom edge.
 
 ```tsx title="Virtual keyboard aware drawer"
-<Drawer.Root virtualKeyboardAware>
+<Drawer.Root>
+  <Drawer.VirtualKeyboardProvider>{/* ... */}</Drawer.VirtualKeyboardProvider>
+</Drawer.Root>
 ```
 
 <DemoDrawerVirtualKeyboardAware />
