+++ docs/src/app/(docs)/react/components/drawer/types.md
@@ -29,7 +29,6 @@ Doesn't render its own HTML element.
 | snapToSequentialPoints  | `boolean`                                                                                               | `false`  | Disables velocity-based snap skipping so drag distance determines the next snap point.                                                                                                                                                                                                                                                                                                                                                            |
 | swipeDirection          | `DrawerSwipeDirection`                                                                                  | `'down'` | The swipe direction used to dismiss the drawer.                                                                                                                                                                                                                                                                                                                                                                                                   |
 | triggerId               | `string \| null`                                                                                        | -        | ID of the trigger that the drawer is associated with.&#xA;This is useful in conjunction with the `open` prop to create a controlled drawer.&#xA;There's no need to specify this prop when the drawer is uncontrolled (that is, when the `open` prop is not set).                                                                                                                                                                                  |
-| virtualKeyboardAware    | `boolean`                                                                                               | `false`  | Enables virtual-keyboard-aware form handling and visible viewport sizing for software keyboards.                                                                                                                                                                                                                                                                                                                                                  |
 | children                | `React.ReactNode \| PayloadChildRenderFunction<Payload>`                                                | -        | The content of the drawer.                                                                                                                                                                                                                                                                                                                                                                                                                        |
 
 ### Root.Props
@@ -617,6 +616,27 @@ type DrawerSwipeAreaState = {
 };
 ```
 
+### VirtualKeyboardProvider
+
+Enables visible viewport sizing and keyboard-aware focus behavior for software keyboards.
+Doesn't render its own HTML element.
+
+**VirtualKeyboardProvider Props:**
+
+| Prop     | Type              | Default | Description |
+| :------- | :---------------- | :------ | :---------- |
+| children | `React.ReactNode` | -       | -           |
+
+### VirtualKeyboardProvider.Props
+
+Re-export of [VirtualKeyboardProvider](#virtualkeyboardprovider) props.
+
+### VirtualKeyboardProvider.State
+
+```typescript
+type DrawerVirtualKeyboardProviderState = {};
+```
+
 ## External Types
 
 ### InteractionType
@@ -670,10 +690,11 @@ type SwipeDirection = 'up' | 'down' | 'left' | 'right';
 - `Drawer.SwipeArea`: `Drawer.SwipeArea`, `Drawer.SwipeArea.Props`, `Drawer.SwipeArea.State`
 - `Drawer.Title`: `Drawer.Title`, `Drawer.Title.Props`, `Drawer.Title.State`
 - `Drawer.Trigger`: `Drawer.Trigger`, `Drawer.Trigger.Props`, `Drawer.Trigger.State`
+- `Drawer.VirtualKeyboardProvider`: `Drawer.VirtualKeyboardProvider`, `Drawer.VirtualKeyboardProvider.State`, `Drawer.VirtualKeyboardProvider.Props`
 - `Drawer.Viewport`: `Drawer.Viewport`, `Drawer.Viewport.Props`, `Drawer.Viewport.State`
 - `Drawer.createHandle`
 - `Drawer.Handle`
-- `Default`: `DrawerRootState`, `DrawerRootProps`, `DrawerRootActions`, `DrawerRootChangeEventReason`, `DrawerRootChangeEventDetails`, `DrawerRootSnapPointChangeEventReason`, `DrawerRootSnapPointChangeEventDetails`, `DrawerProviderState`, `DrawerProviderProps`, `DrawerIndentState`, `DrawerIndentProps`, `DrawerIndentBackgroundState`, `DrawerIndentBackgroundProps`, `DrawerTriggerProps`, `DrawerTriggerState`, `DrawerPortalState`, `DrawerPortalProps`, `DrawerPopupProps`, `DrawerPopupState`, `DrawerSwipeAreaProps`, `DrawerSwipeAreaState`, `DrawerContentProps`, `DrawerContentState`, `DrawerBackdropProps`, `DrawerBackdropState`, `DrawerViewportState`, `DrawerViewportProps`, `DrawerTitleProps`, `DrawerTitleState`, `DrawerDescriptionProps`, `DrawerDescriptionState`, `DrawerCloseProps`, `DrawerCloseState`
+- `Default`: `DrawerRootState`, `DrawerRootProps`, `DrawerRootActions`, `DrawerRootChangeEventReason`, `DrawerRootChangeEventDetails`, `DrawerRootSnapPointChangeEventReason`, `DrawerRootSnapPointChangeEventDetails`, `DrawerProviderState`, `DrawerProviderProps`, `DrawerIndentState`, `DrawerIndentProps`, `DrawerIndentBackgroundState`, `DrawerIndentBackgroundProps`, `DrawerTriggerProps`, `DrawerTriggerState`, `DrawerPortalState`, `DrawerPortalProps`, `DrawerPopupProps`, `DrawerPopupState`, `DrawerSwipeAreaProps`, `DrawerSwipeAreaState`, `DrawerContentProps`, `DrawerContentState`, `DrawerBackdropProps`, `DrawerBackdropState`, `DrawerViewportState`, `DrawerViewportProps`, `DrawerTitleProps`, `DrawerTitleState`, `DrawerDescriptionProps`, `DrawerDescriptionState`, `DrawerCloseProps`, `DrawerCloseState`, `DrawerVirtualKeyboardProviderState`, `DrawerVirtualKeyboardProviderProps`
 
 ## Canonical Types
 
@@ -710,5 +731,7 @@ Maps `Canonical`: `Alias` — Use Canonical when its namespace is already import
 - `Drawer.Title.State`: `DrawerTitleState`
 - `Drawer.Trigger.Props`: `DrawerTriggerProps`
 - `Drawer.Trigger.State`: `DrawerTriggerState`
+- `Drawer.VirtualKeyboardProvider.State`: `DrawerVirtualKeyboardProviderState`
+- `Drawer.VirtualKeyboardProvider.Props`: `DrawerVirtualKeyboardProviderProps`
 - `Drawer.Viewport.Props`: `DrawerViewportProps`
 - `Drawer.Viewport.State`: `DrawerViewportState`
