+++ docs/src/app/(docs)/react/components/page.mdx
@@ -734,7 +734,7 @@ A panel that slides in from the edge of the screen.
     - Handle
 - Exports:
   - Drawer - Root
-    - Props: actionsRef, children, defaultOpen, defaultSnapPoint, defaultTriggerId, disablePointerDismissal, handle, modal, onOpenChange, onOpenChangeComplete, onSnapPointChange, open, snapPoint, snapPoints, snapToSequentialPoints, swipeDirection, triggerId, virtualKeyboardAware
+    - Props: actionsRef, children, defaultOpen, defaultSnapPoint, defaultTriggerId, disablePointerDismissal, handle, modal, onOpenChange, onOpenChangeComplete, onSnapPointChange, open, snapPoint, snapPoints, snapToSequentialPoints, swipeDirection, triggerId
   - Drawer - Provider
     - Props: children
   - Drawer - Trigger
@@ -770,7 +770,9 @@ A panel that slides in from the edge of the screen.
   - Drawer - SwipeArea
     - Props: className, disabled, render, style, swipeDirection
     - Data Attributes: data-closed, data-disabled, data-open, data-swipe-direction, data-swiping
-- Types: Drawer.Backdrop.Props, Drawer.Backdrop.State, Drawer.Close.Props, Drawer.Close.State, Drawer.Content.Props, Drawer.Content.State, Drawer.Description.Props, Drawer.Description.State, Drawer.Indent.Props, Drawer.Indent.State, Drawer.IndentBackground.Props, Drawer.IndentBackground.State, Drawer.Popup.Props, Drawer.Popup.State, Drawer.Portal.Props, Drawer.Portal.State, Drawer.Provider.Props, Drawer.Provider.State, Drawer.Root.Actions, Drawer.Root.ChangeEventDetails, Drawer.Root.ChangeEventReason, Drawer.Root.Props, Drawer.Root.SnapPoint, Drawer.Root.SnapPointChangeEventDetails, Drawer.Root.SnapPointChangeEventReason, Drawer.Root.State, Drawer.SwipeArea.Props, Drawer.SwipeArea.State, Drawer.Title.Props, Drawer.Title.State, Drawer.Trigger.Props, Drawer.Trigger.State, Drawer.Viewport.Props, Drawer.Viewport.State
+  - Drawer - VirtualKeyboardProvider
+    - Props: children
+- Types: Drawer.Backdrop.Props, Drawer.Backdrop.State, Drawer.Close.Props, Drawer.Close.State, Drawer.Content.Props, Drawer.Content.State, Drawer.Description.Props, Drawer.Description.State, Drawer.Indent.Props, Drawer.Indent.State, Drawer.IndentBackground.Props, Drawer.IndentBackground.State, Drawer.Popup.Props, Drawer.Popup.State, Drawer.Portal.Props, Drawer.Portal.State, Drawer.Provider.Props, Drawer.Provider.State, Drawer.Root.Actions, Drawer.Root.ChangeEventDetails, Drawer.Root.ChangeEventReason, Drawer.Root.Props, Drawer.Root.SnapPoint, Drawer.Root.SnapPointChangeEventDetails, Drawer.Root.SnapPointChangeEventReason, Drawer.Root.State, Drawer.SwipeArea.Props, Drawer.SwipeArea.State, Drawer.Title.Props, Drawer.Title.State, Drawer.Trigger.Props, Drawer.Trigger.State, Drawer.Viewport.Props, Drawer.Viewport.State, Drawer.VirtualKeyboardProvider.Props, Drawer.VirtualKeyboardProvider.State
 
 </details>
