+++ packages/react/src/drawer/index.parts.ts
@@ -11,6 +11,7 @@ export { DrawerRoot as Root } from './root/DrawerRoot';
 export { DrawerSwipeArea as SwipeArea } from './swipe-area/DrawerSwipeArea';
 export { DrawerTitle as Title } from './title/DrawerTitle';
 export { DrawerTrigger as Trigger } from './trigger/DrawerTrigger';
+export { DrawerVirtualKeyboardProvider as VirtualKeyboardProvider } from './virtual-keyboard/DrawerVirtualKeyboardProvider';
 export { DrawerViewport as Viewport } from './viewport/DrawerViewport';
 export {
   createDialogHandle as createHandle,
