+++ packages/react/src/drawer/index.ts
@@ -14,3 +14,4 @@ export type * from './viewport/DrawerViewport';
 export type * from './title/DrawerTitle';
 export type * from './description/DrawerDescription';
 export type * from './close/DrawerClose';
+export type * from './virtual-keyboard/DrawerVirtualKeyboardProvider';
