+++ packages/react/src/drawer/root/DrawerRoot.tsx
@@ -44,7 +44,6 @@ export function DrawerRoot<Payload = unknown>(props: DrawerRoot.Props<Payload>) 
     triggerId: triggerIdProp,
     defaultTriggerId: defaultTriggerIdProp = null,
     swipeDirection = 'down',
-    virtualKeyboardAware = false,
     snapToSequentialPoints = false,
     snapPoints,
     snapPoint: snapPointProp,
@@ -173,7 +172,6 @@ export function DrawerRoot<Payload = unknown>(props: DrawerRoot.Props<Payload>) 
   const contextValue: DrawerRootContext = React.useMemo(
     () => ({
       swipeDirection,
-      virtualKeyboardAware,
       snapToSequentialPoints,
       snapPoints,
       activeSnapPoint: resolvedActiveSnapPoint,
@@ -213,7 +211,6 @@ export function DrawerRoot<Payload = unknown>(props: DrawerRoot.Props<Payload>) 
       snapPoints,
       snapToSequentialPoints,
       swipeDirection,
-      virtualKeyboardAware,
     ],
   );
 
@@ -323,11 +320,6 @@ export interface DrawerRootProps<Payload = unknown> {
    * @default 'down'
    */
   swipeDirection?: DrawerSwipeDirection | undefined;
-  /**
-   * Enables virtual-keyboard-aware form handling and visible viewport sizing for software keyboards.
-   * @default false
-   */
-  virtualKeyboardAware?: boolean | undefined;
   /**
    * Snap points used to position the drawer.
    * Use numbers between 0 and 1 to represent fractions of the viewport height,
