+++ packages/react/src/drawer/virtual-keyboard/DrawerVirtualKeyboardContext.tsx
@@ -1,0 +1,30 @@
+'use client';
+import * as React from 'react';
+
+export interface DrawerVirtualKeyboardContext {
+  availableHeight: number | null;
+  keyboardInset: number;
+  onTouchStart: (event: React.TouchEvent<Element>) => void;
+  onTouchEnd: (event: React.TouchEvent<Element>) => boolean;
+  onTouchCancel: () => void;
+}
+
+export const DrawerVirtualKeyboardContext = React.createContext<
+  DrawerVirtualKeyboardContext | undefined
+>(undefined);
+
+export function useDrawerVirtualKeyboardContext(optional?: false): DrawerVirtualKeyboardContext;
+export function useDrawerVirtualKeyboardContext(
+  optional: true,
+): DrawerVirtualKeyboardContext | undefined;
+export function useDrawerVirtualKeyboardContext(optional?: boolean) {
+  const drawerVirtualKeyboardContext = React.useContext(DrawerVirtualKeyboardContext);
+
+  if (optional === false && drawerVirtualKeyboardContext === undefined) {
+    throw new Error(
+      'Base UI: DrawerVirtualKeyboardContext is missing. The virtual keyboard provider must wrap the drawer parts that use it.',
+    );
+  }
+
+  return drawerVirtualKeyboardContext;
+}
